Version 0.16 | Algorithm revision and Optimization

+++ app/src/main/java/com/example/bubblproject/CreateTaskActivity.java
@@ -81,7 +81,7 @@ public class CreateTaskActivity extends AppCompatActivity {
                 }
                 if(priorityText.getText().toString() != "") {
                     try {
-                        task.setTaskPriority(Integer.parseInt(priorityText.getText().toString()));
+                        task.setTaskPriority(Double.parseDouble(priorityText.getText().toString()));
                     }
                     catch (NumberFormatException e){}
                 }
